+++ lib/contact.ts
@@ -3,13 +3,13 @@
 const binding = require('../binding')
 const debug = require('debug')('deltachat:node:contact')
 
+interface NativeContact {}
 /**
  * Wrapper around dc_contact_t*
  */
-class Contact {
-  constructor (dc_contact) {
+export class Contact {
+  constructor (public dc_contact:NativeContact) {
     debug('Contact constructor')
-    this.dc_contact = dc_contact
   }
 
   toJson () {
@@ -28,35 +28,35 @@ class Contact {
     }
   }
 
-  getAddress () {
+  getAddress ():string {
     return binding.dcn_contact_get_addr(this.dc_contact)
   }
 
-  getColor () {
+  getColor ():number {
     return binding.dcn_contact_get_color(this.dc_contact)
   }
 
-  getDisplayName () {
+  getDisplayName ():string {
     return binding.dcn_contact_get_display_name(this.dc_contact)
   }
 
-  getFirstName () {
+  getFirstName ():string{
     return binding.dcn_contact_get_first_name(this.dc_contact)
   }
 
-  getId () {
+  getId ():number{
     return binding.dcn_contact_get_id(this.dc_contact)
   }
 
-  getName () {
+  getName ():string {
     return binding.dcn_contact_get_name(this.dc_contact)
   }
 
-  getNameAndAddress () {
+  getNameAndAddress ():string {
     return binding.dcn_contact_get_name_n_addr(this.dc_contact)
   }
 
-  getProfileImage () {
+  getProfileImage ():string {
     return binding.dcn_contact_get_profile_image(this.dc_contact)
   }
 
@@ -68,5 +68,3 @@ class Contact {
     return Boolean(binding.dcn_contact_is_verified(this.dc_contact))
   }
 }
-
-module.exports = Contact